requirements.txt
beautifulsoup4==4.10.0
bitarray==2.7.5
hdwallet==2.2.1
openpyxl==3.1.2
plyvel_wheels==1.3.0
protobuf==3.12.4
pycryptodome==3.18.0
python_docx==0.8.11